move footer's logic into Presence-footer
